+++ packages/f-card/CHANGELOG.md
@@ -5,13 +5,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 Latest (to be added to next release)
 ------------------------------
-*October 23, 2020*
+*October 26, 2020*
 
 ### Added
 - Stylelint added to lint styling on build.
 - data-test IDs to Card Component 
 - Test for Card
 - Card Component-Object 
+
+### Changed
+- 'jet' theme instead of 'je'
 
 
 v0.7.0